AI Fix: AI fix: Function `addNumbers(int a, int b)` is vulnerable to signed integer overflow when inputs exceed the `int` range. This is undefined behavior in C++. Consider using a wider type such as `long long`, adding overflow checks, or using `std::int64_t` if the domain requires it.

+++ Add.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-int addNumbers(int a, int b) {
+long long addNumbers(long long a, long long b) {
     // Critical issue: Integer overflow can occur here
     return a + b;
 }
@@ -12,7 +12,7 @@ int main() {
     cout << "Enter two numbers: ";
     cin >> x >> y;
 
-    int result = addNumbers(x, y);
+    long long result = addNumbers(x, y);
 
     cout << "Sum = " << result << endl;
